Move printcolor() into color.py file
