+++ src/sdsc/cli.py
@@ -79,22 +79,3 @@ def parseargs(cliargs=None):
     parser.add_argument('outputfile', nargs="?")
 
     return parser.parse_args(args=cliargs)
-
-
-def printcolor(message, messagetype=None):
-    """Print a colored message
-
-    :param str message: Message
-    :param str messagetype: Type (none => green, 'error' => red, 'debug' => blue)
-    """
-    if sys.stdout.isatty():
-        if messagetype == 'error':
-            print('\033[0;31m' + message + '\033[0m', file=sys.stderr)
-        elif messagetype == 'debug':
-            print('\033[0;36m' + message + '\033[0m', file=sys.stderr)
-        else:
-            print('\033[0;32m' + message + '\033[0m', file=sys.stdout)
-    elif messagetype == 'error' or messagetype == 'debug':
-        print(message, file=sys.stderr)
-    else:
-        print(message)
